Module Environment

String-indexed environments.

This module is a thin wrapper around Stdlib.Map specialised to string keys. It exposes the full Stdlib.Map.S interface under the same types and adds a couple of helpers that are commonly needed when manipulating compiler-style environments. Unless stated otherwise, all complexity guarantees and semantics are identical to those of the underlying Stdlib.Map implementation.

include Stdlib.Map.S with type key = string

type key = string

type !+'a t

val empty : 'a t

val add : key -> 'a -> 'a t -> 'a t

val add_to_list : key -> 'a -> 'a list t -> 'a list t

val update : key -> ('a option -> 'a option) -> 'a t -> 'a t

val singleton : key -> 'a -> 'a t

val remove : key -> 'a t -> 'a t

val union : (key -> 'a -> 'a -> 'a option) -> 'a t -> 'a t -> 'a t

val cardinal : 'a t -> int

val bindings : 'a t -> (key * 'a) list

val min_binding : 'a t -> key * 'a

val min_binding_opt : 'a t -> (key * 'a) option

val max_binding : 'a t -> key * 'a

val max_binding_opt : 'a t -> (key * 'a) option

val choose : 'a t -> key * 'a

val choose_opt : 'a t -> (key * 'a) option

val find : key -> 'a t -> 'a

val find_opt : key -> 'a t -> 'a option

val find_first : (key -> bool) -> 'a t -> key * 'a

val find_first_opt : (key -> bool) -> 'a t -> (key * 'a) option

val find_last : (key -> bool) -> 'a t -> key * 'a

val find_last_opt : (key -> bool) -> 'a t -> (key * 'a) option

val iter : (key -> 'a -> unit) -> 'a t -> unit

val fold : (key -> 'a -> 'acc -> 'acc) -> 'a t -> 'acc -> 'acc

val map : ('a -> 'b) -> 'a t -> 'b t

val mapi : (key -> 'a -> 'b) -> 'a t -> 'b t

val filter : (key -> 'a -> bool) -> 'a t -> 'a t

val filter_map : (key -> 'a -> 'b option) -> 'a t -> 'b t

val partition : (key -> 'a -> bool) -> 'a t -> 'a t * 'a t

val split : key -> 'a t -> 'a t * 'a option * 'a t

val is_empty : 'a t -> bool

val mem : key -> 'a t -> bool

val equal : ('a -> 'a -> bool) -> 'a t -> 'a t -> bool

val compare : ('a -> 'a -> int) -> 'a t -> 'a t -> int

val for_all : (key -> 'a -> bool) -> 'a t -> bool

val exists : (key -> 'a -> bool) -> 'a t -> bool

val to_list : 'a t -> (key * 'a) list

val to_seq : 'a t -> (key * 'a) Stdlib.Seq.t

val to_rev_seq : 'a t -> (key * 'a) Stdlib.Seq.t

val to_seq_from : key -> 'a t -> (key * 'a) Stdlib.Seq.t

val add_seq : (key * 'a) Stdlib.Seq.t -> 'a t -> 'a t

val of_seq : (key * 'a) Stdlib.Seq.t -> 'a t

val of_list : (string * 'a) list -> 'a t

of_list bindings builds an environment initialised with the bindings in bindings.

If the list contains several pairs with the same key, the *last* one wins – i.e. the resulting environment is the same as if the pairs were folded from left to right with add.

Example creating an environment with two bindings:

 let env = of_list [ "x", 1; "y", 2 ] in
   find_opt "x" env = Some 1 

val merge : 'a t -> 'a t -> 'a t

merge lhs rhs is the left-biased union of two environments.

All bindings of lhs are kept. For every key that is present *only* in rhs, the corresponding binding is inserted. If a key exists in both maps, the value from lhs is retained – in other words, lhs “wins”.

Example keeping the value from the left map:

  let lhs = of_list [ "x", 1 ] in
  let rhs = of_list [ "x", 0; "y", 2 ] in
  merge lhs rhs |> find_opt "x" = Some 1  (* value from [lhs] *)