Module Filter_file

val filter_lines : input_file:Base.string -> output_file:Base.string -> condition:(Base.string -> bool) -> Base.unit

* filter_lines is a function that filters lines from an input file and writes the filtered lines to an output file. *

parameter input_file
The file to read lines from.

parameter output_file
The file to write filtered lines to.

parameter condition
The condition to filter lines by.