Module Mcp_types.Capability

Partial representation of the capabilities object exchanged during session-initialisation.

Only the keys currently used by the runtime are modelled. All booleans are optional so that the structure remains backward compatible with richer server payloads.

type list_changed = bool

Convenience alias matching the MCP schema (camel-case → snake_case).

type tools_capability = {

list_changed : list_changed option;

}

Capabilities under the tools key.

type prompts_capability = {

list_changed : list_changed option;

}

Capabilities under the prompts key.

type resources_capability = {

subscribe : bool option;
(*
Whether streaming updates are supported.
*)

list_changed : list_changed option;

}

val jsonaf_of_resources_capability : resources_capability -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val resources_capability_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> resources_capability

val sexp_of_resources_capability : resources_capability -> Sexplib0.Sexp.t

val resources_capability_of_sexp : Sexplib0.Sexp.t -> resources_capability

val bin_shape_resources_capability : Bin_prot.Shape.t

val bin_size_resources_capability : resources_capability Bin_prot.Size.sizer

val bin_write_resources_capability : resources_capability Bin_prot.Write.writer

val bin_writer_resources_capability : resources_capability Bin_prot.Type_class.writer

val bin_read_resources_capability : resources_capability Bin_prot.Read.reader

val __bin_read_resources_capability__ : (int -> resources_capability) Bin_prot.Read.reader

val bin_reader_resources_capability : resources_capability Bin_prot.Type_class.reader

val bin_resources_capability : resources_capability Bin_prot.Type_class.t

type t = {

tools : tools_capability option;

prompts : prompts_capability option;

resources : resources_capability option;

}

include Ppx_jsonaf_conv_lib.Jsonafable.S with type t := t

val t_of_jsonaf : Jsonaf_kernel__.Type.t -> t

val jsonaf_of_t : t -> Jsonaf_kernel__.Type.t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

include Bin_prot.Binable.S with type t := t

val bin_size_t : t Bin_prot__.Size.sizer

val bin_write_t : t Bin_prot__.Write.writer

val bin_read_t : t Bin_prot__.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot__.Read.reader

val bin_shape_t : Bin_prot__.Shape.t

val bin_writer_t : t Bin_prot__.Type_class.writer

val bin_reader_t : t Bin_prot__.Type_class.reader

val bin_t : t Bin_prot__.Type_class.t