Module Mcp_types.Jsonrpc

module Id : sig ... end

type error_obj = {

code : int;
(*
Error code defined by JSON-RPC/MCP.
*)

message : string;
(*
Human-readable, non-localized message.
*)

data : Jsonaf.t option;
(*
Optional structured payload.
*)

}

val jsonaf_of_error_obj : error_obj -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val error_obj_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> error_obj

val sexp_of_error_obj : error_obj -> Sexplib0.Sexp.t

val error_obj_of_sexp : Sexplib0.Sexp.t -> error_obj

val bin_shape_error_obj : Bin_prot.Shape.t

val bin_size_error_obj : error_obj Bin_prot.Size.sizer

val bin_write_error_obj : error_obj Bin_prot.Write.writer

val bin_writer_error_obj : error_obj Bin_prot.Type_class.writer

val bin_read_error_obj : error_obj Bin_prot.Read.reader

val __bin_read_error_obj__ : (int -> error_obj) Bin_prot.Read.reader

val bin_reader_error_obj : error_obj Bin_prot.Type_class.reader

val bin_error_obj : error_obj Bin_prot.Type_class.t

type request = {

jsonrpc : string;
(*
Must be the literal "2.0".
*)

id : Id.t;
(*
Correlates request ↔ response.
*)

method_ : string;
(*
RPC method name.
*)

params : Jsonaf.t option;
(*
Optional positional/object params.
*)

}

val jsonaf_of_request : request -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val request_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> request

val sexp_of_request : request -> Sexplib0.Sexp.t

val request_of_sexp : Sexplib0.Sexp.t -> request

val bin_shape_request : Bin_prot.Shape.t

val bin_size_request : request Bin_prot.Size.sizer

val bin_write_request : request Bin_prot.Write.writer

val bin_writer_request : request Bin_prot.Type_class.writer

val bin_read_request : request Bin_prot.Read.reader

val __bin_read_request__ : (int -> request) Bin_prot.Read.reader

val bin_reader_request : request Bin_prot.Type_class.reader

val bin_request : request Bin_prot.Type_class.t

type response = {

jsonrpc : string;
(*
Literal "2.0".
*)

id : Id.t;
(*
Mirrors the request ID.
*)

result : Jsonaf.t option;
(*
Present on success.
*)

error : error_obj option;
(*
Present on failure. result is then None.
*)

}

val jsonaf_of_response : response -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val response_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> response

val sexp_of_response : response -> Sexplib0.Sexp.t

val response_of_sexp : Sexplib0.Sexp.t -> response

val bin_shape_response : Bin_prot.Shape.t

val bin_size_response : response Bin_prot.Size.sizer

val bin_write_response : response Bin_prot.Write.writer

val bin_writer_response : response Bin_prot.Type_class.writer

val bin_read_response : response Bin_prot.Read.reader

val __bin_read_response__ : (int -> response) Bin_prot.Read.reader

val bin_reader_response : response Bin_prot.Type_class.reader

val bin_response : response Bin_prot.Type_class.t

type notification = {

jsonrpc : string;
(*
"2.0".
*)

method_ : string;
(*
Notification name.
*)

params : Jsonaf.t option;
(*
Optional payload.
*)

}

val jsonaf_of_notification : notification -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val notification_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> notification

val sexp_of_notification : notification -> Sexplib0.Sexp.t

val notification_of_sexp : Sexplib0.Sexp.t -> notification

val bin_shape_notification : Bin_prot.Shape.t

val bin_size_notification : notification Bin_prot.Size.sizer

val bin_write_notification : notification Bin_prot.Write.writer

val bin_writer_notification : notification Bin_prot.Type_class.writer

val bin_read_notification : notification Bin_prot.Read.reader

val __bin_read_notification__ : (int -> notification) Bin_prot.Read.reader

val bin_reader_notification : notification Bin_prot.Type_class.reader

val bin_notification : notification Bin_prot.Type_class.t

val make_request : ?params:Jsonaf.t -> id:Id.t -> method_:string -> unit -> request

make_request ?params ~id ~method_ () constructs a well-formed JSON-RPC request envelope with protocol version set to "2.0".

val ok : id:Id.t -> Jsonaf.t -> response

ok ~id payload wraps payload in a successful response envelope.

val error : id:Id.t -> code:int -> message:string -> ?data:Jsonaf.t -> unit -> response

error ~id ~code ~message ?data () constructs an error response.

val notify : ?params:Jsonaf.t -> method_:string -> unit -> notification

notify ?params ~method_ () creates a JSON-RPC notification (no id).