Module Mcp_types.Resource

Metadata for a single resource returned by "resources/list".

Only the fields actively required by the client are represented. The server is free to attach additional keys; they will be ignored during deserialisation.

type t = {

uri : string;
(*
Globally unique resource URI.
*)

name : string;
(*
Display name.
*)

description : string option;
(*
Optional free-form summary.
*)

mime_type : string option;
(*
Media type (e.g. "text/plain", "image/png").
*)

size : int option;
(*
Size in bytes, when known.
*)

}

include Ppx_jsonaf_conv_lib.Jsonafable.S with type t := t

val t_of_jsonaf : Jsonaf_kernel__.Type.t -> t

val jsonaf_of_t : t -> Jsonaf.t

Re-export the auto-generated converter so that callers do not have to depend on Jsonaf_conv.