Module Oauth2_types.Client_registration

type t = {

client_id : string;

client_secret : string option;

}

val t_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> t

val jsonaf_of_t : t -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

Client credentials issued by the authorisation server during *dynamic client registration*.

The record mirrors the minimal fields required by RFC&nbsp;7591 3.2. Applications usually persist the returned pair so that subsequent authorisation requests can authenticate as that client.