Module Oauth2_types.Token

type t = {

access_token : string;

token_type : string;

expires_in : int;

refresh_token : string option;

scope : string option;

obtained_at : float;

}

val t_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> t

val jsonaf_of_t : t -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val is_expired : t -> bool

is_expired t returns true if t will expire within the next minute.

A small *safety window* (60&nbsp;s) is subtracted from expires_in so that callers can refresh the token a bit early and avoid race conditions where a request is dispatched with a token that expires en-route.