Module Prompt.Chat_markdown

type function_call = {

name : string;

arguments : string;

}

val jsonaf_of_function_call : function_call -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val function_call_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> function_call

val sexp_of_function_call : function_call -> Sexplib0.Sexp.t

val function_call_of_sexp : Sexplib0.Sexp.t -> function_call

val hash_fold_function_call : Ppx_hash_lib.Std.Hash.state -> function_call -> Ppx_hash_lib.Std.Hash.state

val hash_function_call : function_call -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_function_call : Bin_prot.Shape.t

val bin_size_function_call : function_call Bin_prot.Size.sizer

val bin_write_function_call : function_call Bin_prot.Write.writer

val bin_writer_function_call : function_call Bin_prot.Type_class.writer

val bin_read_function_call : function_call Bin_prot.Read.reader

val __bin_read_function_call__ : (int -> function_call) Bin_prot.Read.reader

val bin_reader_function_call : function_call Bin_prot.Type_class.reader

val bin_function_call : function_call Bin_prot.Type_class.t

val compare_function_call : function_call -> function_call -> int

type tool_call = {

id : string;

function_ : function_call;

}

val jsonaf_of_tool_call : tool_call -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val tool_call_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> tool_call

val sexp_of_tool_call : tool_call -> Sexplib0.Sexp.t

val tool_call_of_sexp : Sexplib0.Sexp.t -> tool_call

val hash_fold_tool_call : Ppx_hash_lib.Std.Hash.state -> tool_call -> Ppx_hash_lib.Std.Hash.state

val hash_tool_call : tool_call -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_tool_call : Bin_prot.Shape.t

val bin_size_tool_call : tool_call Bin_prot.Size.sizer

val bin_write_tool_call : tool_call Bin_prot.Write.writer

val bin_writer_tool_call : tool_call Bin_prot.Type_class.writer

val bin_read_tool_call : tool_call Bin_prot.Read.reader

val __bin_read_tool_call__ : (int -> tool_call) Bin_prot.Read.reader

val bin_reader_tool_call : tool_call Bin_prot.Type_class.reader

val bin_tool_call : tool_call Bin_prot.Type_class.t

val compare_tool_call : tool_call -> tool_call -> int

type image_url = {

url : string;

}

val jsonaf_of_image_url : image_url -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val image_url_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> image_url

val sexp_of_image_url : image_url -> Sexplib0.Sexp.t

val image_url_of_sexp : Sexplib0.Sexp.t -> image_url

val hash_fold_image_url : Ppx_hash_lib.Std.Hash.state -> image_url -> Ppx_hash_lib.Std.Hash.state

val hash_image_url : image_url -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_image_url : Bin_prot.Shape.t

val bin_size_image_url : image_url Bin_prot.Size.sizer

val bin_write_image_url : image_url Bin_prot.Write.writer

val bin_writer_image_url : image_url Bin_prot.Type_class.writer

val bin_read_image_url : image_url Bin_prot.Read.reader

val __bin_read_image_url__ : (int -> image_url) Bin_prot.Read.reader

val bin_reader_image_url : image_url Bin_prot.Type_class.reader

val bin_image_url : image_url Bin_prot.Type_class.t

val compare_image_url : image_url -> image_url -> int

type basic_content_item = {

type_ : string;

text : string option;

image_url : image_url option;

document_url : string option;

is_local : bool;

cleanup_html : bool;

markdown : bool;

}

val jsonaf_of_basic_content_item : basic_content_item -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val basic_content_item_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> basic_content_item

val sexp_of_basic_content_item : basic_content_item -> Sexplib0.Sexp.t

val basic_content_item_of_sexp : Sexplib0.Sexp.t -> basic_content_item

val hash_fold_basic_content_item : Ppx_hash_lib.Std.Hash.state -> basic_content_item -> Ppx_hash_lib.Std.Hash.state

val hash_basic_content_item : basic_content_item -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_basic_content_item : Bin_prot.Shape.t

val bin_size_basic_content_item : basic_content_item Bin_prot.Size.sizer

val bin_write_basic_content_item : basic_content_item Bin_prot.Write.writer

val bin_writer_basic_content_item : basic_content_item Bin_prot.Type_class.writer

val bin_read_basic_content_item : basic_content_item Bin_prot.Read.reader

val __bin_read_basic_content_item__ : (int -> basic_content_item) Bin_prot.Read.reader

val bin_reader_basic_content_item : basic_content_item Bin_prot.Type_class.reader

val bin_basic_content_item : basic_content_item Bin_prot.Type_class.t

val compare_basic_content_item : basic_content_item -> basic_content_item -> int

type agent_content = {

url : string;

is_local : bool;

items : content_item list;

}

and content_item =

| Basic of basic_content_item

| Agent of agent_content

val jsonaf_of_agent_content : agent_content -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val jsonaf_of_content_item : content_item -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val agent_content_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> agent_content

val content_item_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> content_item

val sexp_of_agent_content : agent_content -> Sexplib0.Sexp.t

val sexp_of_content_item : content_item -> Sexplib0.Sexp.t

val agent_content_of_sexp : Sexplib0.Sexp.t -> agent_content

val content_item_of_sexp : Sexplib0.Sexp.t -> content_item

val hash_fold_agent_content : Ppx_hash_lib.Std.Hash.state -> agent_content -> Ppx_hash_lib.Std.Hash.state

val hash_agent_content : agent_content -> Ppx_hash_lib.Std.Hash.hash_value

val hash_fold_content_item : Ppx_hash_lib.Std.Hash.state -> content_item -> Ppx_hash_lib.Std.Hash.state

val hash_content_item : content_item -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_agent_content : Bin_prot.Shape.t

val bin_shape_content_item : Bin_prot.Shape.t

val bin_size_agent_content : agent_content Bin_prot.Size.sizer

val bin_write_agent_content : agent_content Bin_prot.Write.writer

val bin_writer_agent_content : agent_content Bin_prot.Type_class.writer

val bin_size_content_item : content_item Bin_prot.Size.sizer

val bin_write_content_item : content_item Bin_prot.Write.writer

val bin_writer_content_item : content_item Bin_prot.Type_class.writer

val bin_read_agent_content : agent_content Bin_prot.Read.reader

val __bin_read_agent_content__ : (int -> agent_content) Bin_prot.Read.reader

val bin_reader_agent_content : agent_content Bin_prot.Type_class.reader

val bin_read_content_item : content_item Bin_prot.Read.reader

val __bin_read_content_item__ : (int -> content_item) Bin_prot.Read.reader

val bin_reader_content_item : content_item Bin_prot.Type_class.reader

val bin_agent_content : agent_content Bin_prot.Type_class.t

val bin_content_item : content_item Bin_prot.Type_class.t

val compare_agent_content : agent_content -> agent_content -> int

val compare_content_item : content_item -> content_item -> int

type chat_message_content =

| Text of string

| Items of content_item list

val jsonaf_of_chat_message_content : chat_message_content -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val chat_message_content_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> chat_message_content

val sexp_of_chat_message_content : chat_message_content -> Sexplib0.Sexp.t

val chat_message_content_of_sexp : Sexplib0.Sexp.t -> chat_message_content

val hash_fold_chat_message_content : Ppx_hash_lib.Std.Hash.state -> chat_message_content -> Ppx_hash_lib.Std.Hash.state

val hash_chat_message_content : chat_message_content -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_chat_message_content : Bin_prot.Shape.t

val bin_size_chat_message_content : chat_message_content Bin_prot.Size.sizer

val bin_write_chat_message_content : chat_message_content Bin_prot.Write.writer

val bin_writer_chat_message_content : chat_message_content Bin_prot.Type_class.writer

val bin_read_chat_message_content : chat_message_content Bin_prot.Read.reader

val __bin_read_chat_message_content__ : (int -> chat_message_content) Bin_prot.Read.reader

val bin_reader_chat_message_content : chat_message_content Bin_prot.Type_class.reader

val bin_chat_message_content : chat_message_content Bin_prot.Type_class.t

val compare_chat_message_content : chat_message_content -> chat_message_content -> int

type reasoning_summary = {

text : string;

_type : string;

}

val jsonaf_of_reasoning_summary : reasoning_summary -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val reasoning_summary_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> reasoning_summary

val sexp_of_reasoning_summary : reasoning_summary -> Sexplib0.Sexp.t

val reasoning_summary_of_sexp : Sexplib0.Sexp.t -> reasoning_summary

val hash_fold_reasoning_summary : Ppx_hash_lib.Std.Hash.state -> reasoning_summary -> Ppx_hash_lib.Std.Hash.state

val hash_reasoning_summary : reasoning_summary -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_reasoning_summary : Bin_prot.Shape.t

val bin_size_reasoning_summary : reasoning_summary Bin_prot.Size.sizer

val bin_write_reasoning_summary : reasoning_summary Bin_prot.Write.writer

val bin_writer_reasoning_summary : reasoning_summary Bin_prot.Type_class.writer

val bin_read_reasoning_summary : reasoning_summary Bin_prot.Read.reader

val __bin_read_reasoning_summary__ : (int -> reasoning_summary) Bin_prot.Read.reader

val bin_reader_reasoning_summary : reasoning_summary Bin_prot.Type_class.reader

val bin_reasoning_summary : reasoning_summary Bin_prot.Type_class.t

val compare_reasoning_summary : reasoning_summary -> reasoning_summary -> int

type reasoning = {

summary : reasoning_summary list;

id : string;

status : string option;

_type : string;

}

val jsonaf_of_reasoning : reasoning -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val reasoning_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> reasoning

val sexp_of_reasoning : reasoning -> Sexplib0.Sexp.t

val reasoning_of_sexp : Sexplib0.Sexp.t -> reasoning

val hash_fold_reasoning : Ppx_hash_lib.Std.Hash.state -> reasoning -> Ppx_hash_lib.Std.Hash.state

val hash_reasoning : reasoning -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_reasoning : Bin_prot.Shape.t

val bin_size_reasoning : reasoning Bin_prot.Size.sizer

val bin_write_reasoning : reasoning Bin_prot.Write.writer

val bin_writer_reasoning : reasoning Bin_prot.Type_class.writer

val bin_read_reasoning : reasoning Bin_prot.Read.reader

val __bin_read_reasoning__ : (int -> reasoning) Bin_prot.Read.reader

val bin_reader_reasoning : reasoning Bin_prot.Type_class.reader

val bin_reasoning : reasoning Bin_prot.Type_class.t

val compare_reasoning : reasoning -> reasoning -> int

type msg = {

role : string;

content : chat_message_content option;

name : string option;

id : string option;

status : string option;

function_call : function_call option;

tool_call : tool_call option;

tool_call_id : string option;

}

val jsonaf_of_msg : msg -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val msg_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> msg

val sexp_of_msg : msg -> Sexplib0.Sexp.t

val msg_of_sexp : Sexplib0.Sexp.t -> msg

val hash_fold_msg : Ppx_hash_lib.Std.Hash.state -> msg -> Ppx_hash_lib.Std.Hash.state

val hash_msg : msg -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_msg : Bin_prot.Shape.t

val bin_size_msg : msg Bin_prot.Size.sizer

val bin_write_msg : msg Bin_prot.Write.writer

val bin_writer_msg : msg Bin_prot.Type_class.writer

val bin_read_msg : msg Bin_prot.Read.reader

val __bin_read_msg__ : (int -> msg) Bin_prot.Read.reader

val bin_reader_msg : msg Bin_prot.Type_class.reader

val bin_msg : msg Bin_prot.Type_class.t

val compare_msg : msg -> msg -> int

type user_msg = msg

val jsonaf_of_user_msg : user_msg -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val user_msg_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> user_msg

val sexp_of_user_msg : user_msg -> Sexplib0.Sexp.t

val user_msg_of_sexp : Sexplib0.Sexp.t -> user_msg

val hash_fold_user_msg : Ppx_hash_lib.Std.Hash.state -> user_msg -> Ppx_hash_lib.Std.Hash.state

val hash_user_msg : user_msg -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_user_msg : Bin_prot.Shape.t

val bin_size_user_msg : user_msg Bin_prot.Size.sizer

val bin_write_user_msg : user_msg Bin_prot.Write.writer

val bin_writer_user_msg : user_msg Bin_prot.Type_class.writer

val bin_read_user_msg : user_msg Bin_prot.Read.reader

val __bin_read_user_msg__ : (int -> user_msg) Bin_prot.Read.reader

val bin_reader_user_msg : user_msg Bin_prot.Type_class.reader

val bin_user_msg : user_msg Bin_prot.Type_class.t

val compare_user_msg : user_msg -> user_msg -> int

type assistant_msg = msg

val jsonaf_of_assistant_msg : assistant_msg -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val assistant_msg_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> assistant_msg

val sexp_of_assistant_msg : assistant_msg -> Sexplib0.Sexp.t

val assistant_msg_of_sexp : Sexplib0.Sexp.t -> assistant_msg

val hash_fold_assistant_msg : Ppx_hash_lib.Std.Hash.state -> assistant_msg -> Ppx_hash_lib.Std.Hash.state

val hash_assistant_msg : assistant_msg -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_assistant_msg : Bin_prot.Shape.t

val bin_size_assistant_msg : assistant_msg Bin_prot.Size.sizer

val bin_write_assistant_msg : assistant_msg Bin_prot.Write.writer

val bin_writer_assistant_msg : assistant_msg Bin_prot.Type_class.writer

val bin_read_assistant_msg : assistant_msg Bin_prot.Read.reader

val __bin_read_assistant_msg__ : (int -> assistant_msg) Bin_prot.Read.reader

val bin_reader_assistant_msg : assistant_msg Bin_prot.Type_class.reader

val bin_assistant_msg : assistant_msg Bin_prot.Type_class.t

val compare_assistant_msg : assistant_msg -> assistant_msg -> int

type tool_call_msg = msg

val jsonaf_of_tool_call_msg : tool_call_msg -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val tool_call_msg_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> tool_call_msg

val sexp_of_tool_call_msg : tool_call_msg -> Sexplib0.Sexp.t

val tool_call_msg_of_sexp : Sexplib0.Sexp.t -> tool_call_msg

val hash_fold_tool_call_msg : Ppx_hash_lib.Std.Hash.state -> tool_call_msg -> Ppx_hash_lib.Std.Hash.state

val hash_tool_call_msg : tool_call_msg -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_tool_call_msg : Bin_prot.Shape.t

val bin_size_tool_call_msg : tool_call_msg Bin_prot.Size.sizer

val bin_write_tool_call_msg : tool_call_msg Bin_prot.Write.writer

val bin_writer_tool_call_msg : tool_call_msg Bin_prot.Type_class.writer

val bin_read_tool_call_msg : tool_call_msg Bin_prot.Read.reader

val __bin_read_tool_call_msg__ : (int -> tool_call_msg) Bin_prot.Read.reader

val bin_reader_tool_call_msg : tool_call_msg Bin_prot.Type_class.reader

val bin_tool_call_msg : tool_call_msg Bin_prot.Type_class.t

val compare_tool_call_msg : tool_call_msg -> tool_call_msg -> int

type tool_response_msg = msg

val jsonaf_of_tool_response_msg : tool_response_msg -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val tool_response_msg_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> tool_response_msg

val sexp_of_tool_response_msg : tool_response_msg -> Sexplib0.Sexp.t

val tool_response_msg_of_sexp : Sexplib0.Sexp.t -> tool_response_msg

val hash_fold_tool_response_msg : Ppx_hash_lib.Std.Hash.state -> tool_response_msg -> Ppx_hash_lib.Std.Hash.state

val hash_tool_response_msg : tool_response_msg -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_tool_response_msg : Bin_prot.Shape.t

val bin_size_tool_response_msg : tool_response_msg Bin_prot.Size.sizer

val bin_write_tool_response_msg : tool_response_msg Bin_prot.Write.writer

val bin_writer_tool_response_msg : tool_response_msg Bin_prot.Type_class.writer

val bin_read_tool_response_msg : tool_response_msg Bin_prot.Read.reader

val __bin_read_tool_response_msg__ : (int -> tool_response_msg) Bin_prot.Read.reader

val bin_reader_tool_response_msg : tool_response_msg Bin_prot.Type_class.reader

val bin_tool_response_msg : tool_response_msg Bin_prot.Type_class.t

val compare_tool_response_msg : tool_response_msg -> tool_response_msg -> int

type developer_msg = msg

val jsonaf_of_developer_msg : developer_msg -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val developer_msg_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> developer_msg

val sexp_of_developer_msg : developer_msg -> Sexplib0.Sexp.t

val developer_msg_of_sexp : Sexplib0.Sexp.t -> developer_msg

val hash_fold_developer_msg : Ppx_hash_lib.Std.Hash.state -> developer_msg -> Ppx_hash_lib.Std.Hash.state

val hash_developer_msg : developer_msg -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_developer_msg : Bin_prot.Shape.t

val bin_size_developer_msg : developer_msg Bin_prot.Size.sizer

val bin_write_developer_msg : developer_msg Bin_prot.Write.writer

val bin_writer_developer_msg : developer_msg Bin_prot.Type_class.writer

val bin_read_developer_msg : developer_msg Bin_prot.Read.reader

val __bin_read_developer_msg__ : (int -> developer_msg) Bin_prot.Read.reader

val bin_reader_developer_msg : developer_msg Bin_prot.Type_class.reader

val bin_developer_msg : developer_msg Bin_prot.Type_class.t

val compare_developer_msg : developer_msg -> developer_msg -> int

type system_msg = msg

val jsonaf_of_system_msg : system_msg -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val system_msg_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> system_msg

val sexp_of_system_msg : system_msg -> Sexplib0.Sexp.t

val system_msg_of_sexp : Sexplib0.Sexp.t -> system_msg

val hash_fold_system_msg : Ppx_hash_lib.Std.Hash.state -> system_msg -> Ppx_hash_lib.Std.Hash.state

val hash_system_msg : system_msg -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_system_msg : Bin_prot.Shape.t

val bin_size_system_msg : system_msg Bin_prot.Size.sizer

val bin_write_system_msg : system_msg Bin_prot.Write.writer

val bin_writer_system_msg : system_msg Bin_prot.Type_class.writer

val bin_read_system_msg : system_msg Bin_prot.Read.reader

val __bin_read_system_msg__ : (int -> system_msg) Bin_prot.Read.reader

val bin_reader_system_msg : system_msg Bin_prot.Type_class.reader

val bin_system_msg : system_msg Bin_prot.Type_class.t

val compare_system_msg : system_msg -> system_msg -> int

type custom_tool = {

name : string;

description : string option;

command : string;

}

val jsonaf_of_custom_tool : custom_tool -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val custom_tool_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> custom_tool

val sexp_of_custom_tool : custom_tool -> Sexplib0.Sexp.t

val custom_tool_of_sexp : Sexplib0.Sexp.t -> custom_tool

val hash_fold_custom_tool : Ppx_hash_lib.Std.Hash.state -> custom_tool -> Ppx_hash_lib.Std.Hash.state

val hash_custom_tool : custom_tool -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_custom_tool : Bin_prot.Shape.t

val bin_size_custom_tool : custom_tool Bin_prot.Size.sizer

val bin_write_custom_tool : custom_tool Bin_prot.Write.writer

val bin_writer_custom_tool : custom_tool Bin_prot.Type_class.writer

val bin_read_custom_tool : custom_tool Bin_prot.Read.reader

val __bin_read_custom_tool__ : (int -> custom_tool) Bin_prot.Read.reader

val bin_reader_custom_tool : custom_tool Bin_prot.Type_class.reader

val bin_custom_tool : custom_tool Bin_prot.Type_class.t

val compare_custom_tool : custom_tool -> custom_tool -> int

type agent_tool = {

name : string;

description : string option;

agent : string;

is_local : bool;

}

val jsonaf_of_agent_tool : agent_tool -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val agent_tool_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> agent_tool

val sexp_of_agent_tool : agent_tool -> Sexplib0.Sexp.t

val agent_tool_of_sexp : Sexplib0.Sexp.t -> agent_tool

val hash_fold_agent_tool : Ppx_hash_lib.Std.Hash.state -> agent_tool -> Ppx_hash_lib.Std.Hash.state

val hash_agent_tool : agent_tool -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_agent_tool : Bin_prot.Shape.t

val bin_size_agent_tool : agent_tool Bin_prot.Size.sizer

val bin_write_agent_tool : agent_tool Bin_prot.Write.writer

val bin_writer_agent_tool : agent_tool Bin_prot.Type_class.writer

val bin_read_agent_tool : agent_tool Bin_prot.Read.reader

val __bin_read_agent_tool__ : (int -> agent_tool) Bin_prot.Read.reader

val bin_reader_agent_tool : agent_tool Bin_prot.Type_class.reader

val bin_agent_tool : agent_tool Bin_prot.Type_class.t

val compare_agent_tool : agent_tool -> agent_tool -> int

type mcp_tool = {

names : string list option;

description : string option;

mcp_server : string;

strict : bool;

client_id_env : string option;

client_secret_env : string option;

}

val jsonaf_of_mcp_tool : mcp_tool -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val mcp_tool_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> mcp_tool

val sexp_of_mcp_tool : mcp_tool -> Sexplib0.Sexp.t

val mcp_tool_of_sexp : Sexplib0.Sexp.t -> mcp_tool

val hash_fold_mcp_tool : Ppx_hash_lib.Std.Hash.state -> mcp_tool -> Ppx_hash_lib.Std.Hash.state

val hash_mcp_tool : mcp_tool -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_mcp_tool : Bin_prot.Shape.t

val bin_size_mcp_tool : mcp_tool Bin_prot.Size.sizer

val bin_write_mcp_tool : mcp_tool Bin_prot.Write.writer

val bin_writer_mcp_tool : mcp_tool Bin_prot.Type_class.writer

val bin_read_mcp_tool : mcp_tool Bin_prot.Read.reader

val __bin_read_mcp_tool__ : (int -> mcp_tool) Bin_prot.Read.reader

val bin_reader_mcp_tool : mcp_tool Bin_prot.Type_class.reader

val bin_mcp_tool : mcp_tool Bin_prot.Type_class.t

val compare_mcp_tool : mcp_tool -> mcp_tool -> int

type tool =

| Builtin of string

| Custom of custom_tool

| Agent of agent_tool

| Mcp of mcp_tool

val jsonaf_of_tool : tool -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val tool_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> tool

val sexp_of_tool : tool -> Sexplib0.Sexp.t

val tool_of_sexp : Sexplib0.Sexp.t -> tool

val hash_fold_tool : Ppx_hash_lib.Std.Hash.state -> tool -> Ppx_hash_lib.Std.Hash.state

val hash_tool : tool -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_tool : Bin_prot.Shape.t

val bin_size_tool : tool Bin_prot.Size.sizer

val bin_write_tool : tool Bin_prot.Write.writer

val bin_writer_tool : tool Bin_prot.Type_class.writer

val bin_read_tool : tool Bin_prot.Read.reader

val __bin_read_tool__ : (int -> tool) Bin_prot.Read.reader

val bin_reader_tool : tool Bin_prot.Type_class.reader

val bin_tool : tool Bin_prot.Type_class.t

val compare_tool : tool -> tool -> int

type config = {

max_tokens : int option;

model : string option;

reasoning_effort : string option;

temperature : float option;

show_tool_call : bool;

id : string option;

}

val jsonaf_of_config : config -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val config_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> config

val sexp_of_config : config -> Sexplib0.Sexp.t

val config_of_sexp : Sexplib0.Sexp.t -> config

val hash_fold_config : Ppx_hash_lib.Std.Hash.state -> config -> Ppx_hash_lib.Std.Hash.state

val hash_config : config -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_config : Bin_prot.Shape.t

val bin_size_config : config Bin_prot.Size.sizer

val bin_write_config : config Bin_prot.Write.writer

val bin_writer_config : config Bin_prot.Type_class.writer

val bin_read_config : config Bin_prot.Read.reader

val __bin_read_config__ : (int -> config) Bin_prot.Read.reader

val bin_reader_config : config Bin_prot.Type_class.reader

val bin_config : config Bin_prot.Type_class.t

val compare_config : config -> config -> int

type top_level_elements =

| Msg of msg
(*
Legacy <msg/> element (system, developer…)
*)

| Developer of developer_msg

| System of system_msg

| User of user_msg
(*
<user/>
*)

| Assistant of assistant_msg
(*
<assistant/>
*)

| Tool_call of tool_call_msg
(*
<tool_call/>
*)

| Tool_response of tool_response_msg
(*
<tool_response/>
*)

| Config of config

| Reasoning of reasoning

| Tool of tool

val jsonaf_of_top_level_elements : top_level_elements -> Ppx_jsonaf_conv_lib.Jsonaf_kernel.t

val top_level_elements_of_jsonaf : Ppx_jsonaf_conv_lib.Jsonaf_kernel.t -> top_level_elements

val sexp_of_top_level_elements : top_level_elements -> Sexplib0.Sexp.t

val top_level_elements_of_sexp : Sexplib0.Sexp.t -> top_level_elements

val hash_fold_top_level_elements : Ppx_hash_lib.Std.Hash.state -> top_level_elements -> Ppx_hash_lib.Std.Hash.state

val hash_top_level_elements : top_level_elements -> Ppx_hash_lib.Std.Hash.hash_value

val bin_shape_top_level_elements : Bin_prot.Shape.t

val bin_size_top_level_elements : top_level_elements Bin_prot.Size.sizer

val bin_write_top_level_elements : top_level_elements Bin_prot.Write.writer

val bin_writer_top_level_elements : top_level_elements Bin_prot.Type_class.writer

val bin_read_top_level_elements : top_level_elements Bin_prot.Read.reader

val __bin_read_top_level_elements__ : (int -> top_level_elements) Bin_prot.Read.reader

val bin_reader_top_level_elements : top_level_elements Bin_prot.Type_class.reader

val bin_top_level_elements : top_level_elements Bin_prot.Type_class.t

val compare_top_level_elements : top_level_elements -> top_level_elements -> int

val parse_chat_inputs : dir:Eio.Fs.dir_ty Eio.Path.t -> string -> top_level_elements list

parse_chat_inputs ~dir src tokenises, parses and normalises the ChatMarkdown document contained in src. The result is the ordered sequence of top_level_elements that should be sent to the LLM.

– dir is the base directory used to resolve relative paths appearing in `<import src="…"/>`, `<img local="true"/>` or `<agent local="true"/>` tags. The function relies on Io.load_doc under the hood and therefore performs read-only I/O.

The helper expands imports recursively, translates shorthand tags (`<user/>`, `<assistant/>`, `<tool_call/>`, `<tool_response/>`) into dedicated record aliases, and discards elements that are irrelevant for the conversation buffer (e.g. whitespace between top-level nodes).

raises Failure
if the document is malformed or if an imported file cannot be read.