Module Template.Make_Template

The Make_Template functor instantiates a templating module for a given RENDERABLE. Placeholders follow the `{variable

}

` convention.

Parameters

module R : RENDERABLE

Signature

type t

val create : string -> t

create s returns a template whose literal content is s.

The string may contain placeholders such as

 {{user}} 

that correspond to keys produced by R.to_key_value_pairs.

val render : t -> R.t -> string

render tmpl v substitutes every placeholder in tmpl with the value associated with the same key in v. Placeholders that do not have a corresponding key are replaced by the empty string.

Example:

  let module T = Make_Template (struct
    type t = unit
    let to_key_value_pairs () = [ "x", "42" ]
  end) in
  T.render (T.create "value = {{x}}") () = "value = 42"

val to_string : t -> string

to_string tmpl returns the literal representation of tmpl without performing any substitution. This is equivalent to the argument passed to create.