Package lwt.unix

Library lwt.unix

This library exposes the following toplevel modules:

Lwt_bytes Byte arrays

Lwt_config

Lwt_engine Lwt unix main loop engine

Lwt_features

Lwt_fmt Format API for Lwt-powered IOs

Lwt_gc Interaction with the garbage collector

Lwt_io Buffered byte channels

Lwt_main Main loop and event queue

Lwt_preemptive This module allows to mix preemptive threads with Lwt cooperative threads. It maintains an extensible pool of preemptive threads to which you can detach computations.

Lwt_process Process management

Lwt_sys System informations.

Lwt_throttle Rate limiters.

Lwt_timeout Cancelable timeouts.

Lwt_unix Cooperative system calls