Package chatgpt

Sub-indexes

apply_patch

bin_prot_utils

bin_prot_utils_eio

chat_completion

chat_response

chatml

chatml_builtin_modules

chatml_typechecker

definitions

doc

dune_describe

filter_file

functions

github

gpt_function

indexer

io

jsonaf_ext

lru_cache

merlin

ocaml_parser

openai

openai_chat_completion

prompt_template

template

tikitoken

ttl_lru_cache

vector_db

Library chatgpt.apply_patch

The entry point of this library is the module: Apply_patch.

Library chatgpt.bin_prot_utils

The entry point of this library is the module: Bin_prot_utils.

Library chatgpt.bin_prot_utils_eio

The entry point of this library is the module: Bin_prot_utils_eio.

Library chatgpt.chat_completion

The entry point of this library is the module: Chat_completion.

Library chatgpt.chat_response

The entry point of this library is the module: Chat_response.

Library chatgpt.chatml

The entry point of this library is the module: Chatml.

Library chatgpt.chatml_builtin_modules

The entry point of this library is the module: Chatml_builtin_modules.

Library chatgpt.chatml_typechecker

The entry point of this library is the module: Chatml_typechecker.

Library chatgpt.definitions

The entry point of this library is the module: Definitions.

Library chatgpt.doc

The entry point of this library is the module: Doc.

Library chatgpt.dune_describe

The entry point of this library is the module: Dune_describe.

Library chatgpt.filter_file

The entry point of this library is the module: Filter_file.

Library chatgpt.functions

The entry point of this library is the module: Functions.

Library chatgpt.github

The entry point of this library is the module: Github.

Library chatgpt.gpt_function

The entry point of this library is the module: Gpt_function.

Library chatgpt.indexer

The entry point of this library is the module: Indexer.

Library chatgpt.io

The entry point of this library is the module: Io.

Library chatgpt.jsonaf_ext

The entry point of this library is the module: Jsonaf_ext.

Library chatgpt.lru_cache

The entry point of this library is the module: Lru_cache.

Library chatgpt.merlin

The entry point of this library is the module: Merlin.

Library chatgpt.ocaml_parser

The entry point of this library is the module: Ocaml_parser.

Library chatgpt.openai

The entry point of this library is the module: Openai.

Library chatgpt.openai_chat_completion

The entry point of this library is the module: Openai_chat_completion.

Library chatgpt.prompt_template

The entry point of this library is the module: Prompt_template.

Library chatgpt.template

The entry point of this library is the module: Template.

Library chatgpt.tikitoken

The entry point of this library is the module: Tikitoken.

Library chatgpt.ttl_lru_cache

The entry point of this library is the module: Ttl_lru_cache.

Library chatgpt.vector_db

The entry point of this library is the module: Vector_db.